+++ tool/lib/test/unit/parallel.rb
@@ -6,7 +6,7 @@ require "profile_test_all" if ENV.key?('RUBY_TEST_ALL_PROFILE')
 require "tracepointchecker"
 require "zombie_hunter"
 require "iseq_loader_checker"
-require "gc_compact_checker"
+require "gc_checker"
 
 module Test
   module Unit